View now extends GenTraversableOnce.
This should allow views to be used inside flatMap operations.
We assume this is ok because GenTraversaleOnce has NO implementations of any methods, so we
need to always provide our own implementations.

TODO - We should check to see if GenTraversable[E] makes more sense to extend.  That one
has a view issues in terms of what it assumes (e.g. ).